chore(api): migrate admin campaigns list to prescription context
