+++ api/lib/application/organizations/index.js
@@ -138,35 +138,6 @@ const register = async function (server) {
     },
     {
       method: 'GET',
-      path: '/api/admin/organizations/{id}/campaigns',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.organizationId,
-          }),
-        },
-        handler: organizationController.findPaginatedCampaignManagements,
-        tags: ['api', 'organizations'],
-        notes: [
-          'Cette route est restreinte aux administrateurs authentifiés',
-          'Elle retourne toutes les campagnes rattachées à l’organisation.',
-        ],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/organizations/{id}/invitations',
       config: {
         pre: [
