+++ api/src/prescription/campaign/domain/usecases/find-paginated-campaign-managements.js
