+++ api/src/prescription/campaign/infrastructure/serializers/jsonapi/campaign-management-serializer.js
